added travis deployment for terraform

+++ run-terraform.sh
@@ -1,18 +1,21 @@
 #!/bin/bash
 
-cd credentials
-. ./credentials.sh
-cd ..
-
-cd run-dir
-rm *.tf
-cd ..
+if [ ! -d "run-dir" ]; then
+  	mkdir run-dir
+else
+  	cd run-dir
+  	rm *.tf
+  	cd ..
+fi
 
 for DIR in terraform/*; do
+	DIRNAME=$(basename $DIR)
 	for FILE in $DIR/*; do
-		cp $FILE run-dir
+		FILENAME=$(basename $FILE)
+		cp $FILE run-dir/${DIRNAME}_${FILENAME}
 	done
 done
 
+cp bin/terraform run-dir
+
 cp init.tf run-dir
-cd run-dir
